Allow the equal key to increment the date like the plus key
The plus and minus key can be used to adjust the date by one day in each
direction. On many keyboard layouts, the plus key can only be reached by
pressing it together with the shift modifier. In those cases, on most
keyboard layouts one finds the plus being on the same key as the equal
sign.

In the KMyMoney 5.1 series the equal key therefore also incremented the
date by one day. This change brings back that functionality which was
lost during the transition to the new date widget.

BUG: 507964
FIXED-IN: 5.2.2