Fix display of preferred accounts on home page
The home page has two selectable sections: payment accounts and preferred accounts. If the section for preferred accounts was selected, the entries were also listed in the payment accounts section. If it was deselected, the preferred accounts did not show up at all. Actually it should be the other way around: Preferred accounts are shown in their own section but when the section is disabled they should be listed in payment accounts. This change fixes this problem.