Move the account selector of the ledger view to the left side
In case a user has many accounts open in the ledger view, the tab with
the new account selector may become invisible and the user has to shift
the tabs to the left to get the selector into view.

Showing the new account selector on the left most tab resolves this
issue and leaves the + marked tab accessible at program start no matter
how many accounts are displayed.