Make sure hidden columns are really not visible
It has been noticed, that if the application's configuration is lost, the
inital setup hidden columns are not filled with data but they are visible
in the view. Calling the hide/showSection again when the header view is
about to be displayed solves the issue and removes them completely.

(cherry picked from commit 5100d71b00a502568a08e2b2005d9827e31a3e86)