Use correct method to determine top level parent id
In case an account referenced an unknown parent account duing loading,
it needs to be reparented to the corresponding top level account group.
The logic to determine that top level account group did not work as
expected so that an empty account id was used, which caused problems
(crashes) later on. This change now uses the new method to get the right
parent id.

CCBUG: 507416