Focus out on payee widget takes the selected payee from the popup
Hitting TAB in the payee widget only asks to create a new payee if the
name is not on file and the popup is not visible. In case the popup
contains a selected entry, that name is used.

BUG: 508989
FIXED-IN: 5.2.2
(cherry picked from commit 5213e86c5383ac318514b4ef097a2d6dc4f43c29)