Clear the split model before loading another transaction
When using the autofill transaction feature together with the default
category assignment and a transaction was selected from the list of
previous transactions, the default category was added as an additional
split. This change prevents that additional split to be left over.

BUG: 509138
FIXED-IN: 5.2.2
(cherry picked from commit 02f047cebdba8bf814a8bbbfab86112da67e262e)