Update release information in release branch
The release information in the appdata file is used to report what version a flatpak is (e.g. with `flatpak info org.kde.kmymoney`). If it is not up to date in the tarball an older version will be reported. See https://github.com/flathub/org.kde.kmymoney/issues/44 or the release info at https://flathub.org/en/apps/org.kde.kmymoney showing 5.1.0 although the actual version built is 5.1.3