Prevent crash by eliminating lambda slot
The connection needs to be released before the object is destroyed which
otherwise could cause a crash. This can only be done with a normal
method but not with a lambda function. Converting the lambda into a
normal slot resolves the issue.

BUG: 510209
FIXED-IN: 5.2.2
(cherry picked from commit a31e15fc045a5a6de32969a91ade3e9edf61370d)