Fix price display in investment transaction editor
Adjusting the price in the widget is only necessary when the price is
managed per share. It is not needed when entering the total value of all
shares for the transaction.

This amends commit 873dfec0

(cherry picked from commit 6741b0e29e38b5c5c77883c283c32314b72e4c7e)