flatpak: update xmlsec
Download from GitHub so the URL is stable.
Only the current release tarballs are in
https://www.aleksey.com/xmlsec/download/ , the older releases get moved to
https://www.aleksey.com/xmlsec/download/older-releases/ ,
which means the URL for a certain release does not stay the same.
This breaks the build when a new version of xmlsec is released.

(cherry picked from commit dbbc074976d5e7d7600dbef74df871f029ec8f44)