Remove explicit IDs in plugins
Plugin IDs are derived from the filename anyway.

This fixes warnings such as:
kf.coreaddons: The plugin "/home/mark/kde/usr/lib64/plugins/kmymoney_plugins/konlinetasks_sepa.so"
explicitly states an Id in the embedded metadata, which is different from the one derived from
the filename The Id field from the KPlugin object in the metadata should be removed
or:
kf.coreaddons: The plugin "/home/mark/kde/usr/lib64/plugins/kmymoney_plugins/csvimporter.so"
explicitly states an 'Id' in the embedded metadata. This value should be removed, the
resulting pluginId will not be affected by it