Keep column settings in institutions view
The institutions view did not restore user modifications correctly. This
change fixes the problem. The trick is to create the ColumnSelector
object on a view that does not have a model assigned and load the stored
settings during the assignment of the model.

(cherry picked from commit feaa732d212e986c32fb9561b5e8c602531c23ed)