Improve reconciliation performance
When finishing the reconciliation the application needed a long time to
perform the action. This was caused due to cache updates on every
transaction that was marked reconciled. Recalculating the cache at the
end of the operation is enough and speeds up the process significantly.

(cherry picked from commit 7075488482fff6aeab364d75db48bfcbaf4c0344)