Keep the schedule type for loans
Using the new edit feature for scheduled loan transactions changed the
type of the schedule from loan-payment to bill. This resulted in the
automatic calculation of the values for interest and amortization not
being performed correctly in other areas of the application and
incorrect data being displayed.

The change also includes a correction of the data in case a non-loan
type schedule contains automatically calculated split values back to
type loan for erroneous records.

BUG: 513387
FIXED-IN: 5.2.2
(cherry picked from commit 5b5b55c9680523a846257f8849a583bd3bf5c22d)