Allow chaning the memo in multiple transactions at once
When multiple transactions are edited, entering text in the memo widget has no effect. This change fixes the problem so that the data is stored. If only spaces are entered, the memo is cleared. In this case the user will be asked if that is what he wants to do.

BUG: 513948
FIXED-IN: 5.2.2
(cherry picked from commit 56baf19e1e70e96d82942f6ce952f350980b7ddd)